+++ qbraid/runtime/openquantum/device.py
@@ -94,7 +94,10 @@ class OpenQuantumDevice(QuantumDevice):
         if organization_id is None:
             orgs = self.session.get_user_organizations()
             if not orgs:
-                raise QbraidRuntimeError("No organization found for user.")
+                raise QbraidRuntimeError(
+                    "No organization found. Please accept the Open Quantum "
+                    "terms of use to continue."
+                )
             organization_id = orgs[0]["id"]
 
         # Encode QASM